requirements.txt
alembic==1.13.2
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==2.0.12
click==8.1.7
distlib==0.3.8
exceptiongroup==1.2.1
filelock==3.15.4
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
gunicorn==21.1.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.0
pytz==2024.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.19
virtualenv==20.26.3
Werkzeug==2.3.8
zipp==3.19.2
